Generalize code to accept other screens
