+++ pyqt-apps/scripts/sirius-hla-as-cr-control.py
@@ -5,11 +5,16 @@
 import sys
 import argparse as _argparse
 from siriushla.sirius_application import SiriusApplication
-from siriushla.as_cr_control import CryoMain
+from siriushla.as_cr_control import CryoControl
 
 parser = _argparse.ArgumentParser(description="Run Cryogenic Interface.")
+parser.add_argument('screen', type=str, default='All',
+    help='Select the screen of the Cryogenic GUI')
+args = parser.parse_args()
+
 
 app = SiriusApplication()
+
 app.open_window(
-    CryoMain, parent=None)
+    CryoControl, parent=None, screen=args.screen)
 sys.exit(app.exec_())
