+++ pyqt-apps/siriushla/as_cr_control/__init__.py
@@ -1,1 +1,1 @@
-from .main import CryoMain
+from .main import CryoControl
