+++ pyqt-apps/siriushla/as_cr_control/colors.py
@@ -1,0 +1,6 @@
+class COLOR:
+    black = "#000000"
+    dark_red = "#6b1c40"
+    dark_blue = "#325692"
+    blue = "#0000ff"
+    gold = "#808000"
